requirements.txt

streamlit
streamlit-authenticator
PyYAML
pandas
requests
